+++ package-dev/Runtime/SentryInitialization.cs
@@ -20,6 +20,10 @@
 #define SENTRY_NATIVE_SWITCH
 #endif
 
+#if UNITY_SWITCH2
+#define SENTRY_NATIVE_SWITCH2
+#endif
+
 #if UNITY_WEBGL
 #define SENTRY_WEBGL
 #endif
@@ -45,7 +49,7 @@ using Sentry.Unity.MacOS;
 using Sentry.Unity.iOS;
 #elif SENTRY_NATIVE_ANDROID
 using Sentry.Unity.Android;
-#elif SENTRY_NATIVE || SENTRY_NATIVE_SWITCH
+#elif SENTRY_NATIVE || SENTRY_NATIVE_SWITCH || SENTRY_NATIVE_SWITCH2
 using Sentry.Unity.Native;
 #elif SENTRY_WEBGL
 using Sentry.Unity.WebGL;
@@ -108,7 +112,7 @@ namespace Sentry.Unity
             SentryPlatformServices.PlatformConfiguration = SentryNativeCocoa.Configure;
 #elif SENTRY_NATIVE_ANDROID
             SentryPlatformServices.PlatformConfiguration = SentryNativeAndroid.Configure;
-#elif SENTRY_NATIVE_SWITCH
+#elif SENTRY_NATIVE_SWITCH || SENTRY_NATIVE_SWITCH2
             SentryPlatformServices.PlatformConfiguration = SentryNativeSwitch.Configure;
 #elif SENTRY_NATIVE
             SentryPlatformServices.PlatformConfiguration = SentryNative.Configure;